v3.9.6-rc2 minor polish
[features]
* install pacman-contrib if cleanup is enabled and configured to use paccache

[bug fix]
* make certain package install checks fuzzy instead of exact

[aesthetics]
* output when and why missing dependency is being installed

+++ auto-update.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #Auto Update For Manjaro by Lectrode
-vsn="v3.9.6-rc1"; vsndsp="$vsn 2022-07-17"
+vsn="v3.9.6-rc2"; vsndsp="$vsn 2022-07-18"
 #-Downloads and Installs new updates
 #-Depends: coreutils, grep, pacman, pacman-mirrors, iputils
 #-Optional Depends: flatpak, notify-desktop, pikaur, rebuild-detector, wget
@@ -45,8 +45,9 @@ trblqout(){
 
 test_online(){ ping -c 1 "${conf_a[main_testsite_str]}" >/dev/null 2>&1 && return 0; return 1; }
 
-chk_pkgisinst(){ if [[ "$($pcmbin -Qq "$1" 2>/dev/null | grep -m1 -x "$1")" == "$1" ]]; then [[ "$2" = "1" ]] && echo "$1"; return 0; else return 1; fi; }
-chk_pkgisexplicit(){ $pcmbin -Qi "$1" 2>/dev/null|grep -F "Install Reason"|grep -F "Explicitly" >/dev/null && return 0; return 1; }
+chk_pkginst(){ $pcmbin -Q "$1" >/dev/null 2>&1 || return 1; }
+chk_pkginstx(){ [[ "$($pcmbin -Qq "$1" 2>/dev/null | grep -m1 -x "$1")" == "$1" ]] || return 1; }
+chk_pkgexplicit(){ $pcmbin -Qi "$1" 2>/dev/null|grep -F "Install Reason"|grep -F "Explicitly" >/dev/null && return 0; return 1; }
 get_pkgvsn(){ $pcmbin -Q "$1" | grep "$1 " -m1 | cut -d' ' -f2; }
 chk_pkgvsndiff(){ cpvd_t1="$(get_pkgvsn "$1")"; vercmp "${cpvd_t1:-0}" "$2"; unset cpvd_t1; }
 chk_sha256(){ [[ "$(sha256sum "$1" |cut -d ' ' -f 1 |tr -cd '[:alnum:]')" = "$2" ]] && return 0; return 1; }
@@ -102,11 +103,19 @@ if [[ "$1" = "DBPath" ]]; then echo "/var/lib/pacman"; fi
 if [[ "$1" = "CacheDir" ]]; then echo "/var/cache/pacman/pkg"; fi
 }
 
+inst_misspkg(){
+trbl "$2. Installing missing dependency: $1"
+if chk_remoterepo "$1"; then $pcmbin -S --noconfirm "$1"|trbl_t; return; fi
+for h in pikaur apacman; do if [ "${hlpr_a[$h]}" = "1" ]; then $h -S --needed --noconfirm "$1"; break; fi; done
+}
+
 pcln_fol(){
 #$1=num, $2=fol
 [[ -d "$2" ]] || return
 # shellcheck disable=SC2115
 if [[ "$1" = "0" ]]; then rm -rf "$2"/*; return; fi
+if ! paccache --help >/dev/null 2>&1 && chk_remoterepo "pacman-contrib"; then
+    inst_misspkg "pacman-contrib" "Cleanup enabled and configured to use paccache"; fi
 if paccache --help >/dev/null 2>&1; then paccache -rfqk"$1" -c "$2"
 else trbl "$co_y cannot clean $2, paccache not found/functioning"; fi
 }
@@ -150,30 +159,30 @@ then echo crit; else echo norm; fi
 }
 
 manualExplicit(){
-chk_pkgisinst "$1" || return 0
-chk_pkgisexplicit "$1" && return 0
+chk_pkginstx "$1" || return 0
+chk_pkgexplicit "$1" && return 0
 $pcmbin -D --asexplicit "$1"|trbl_t
-chk_pkgisexplicit "$1" && return 0 || return 1
+chk_pkgexplicit "$1" && return 0 || return 1
 }
 
 manualDepend(){
-chk_pkgisinst "$1" || return 0
-chk_pkgisexplicit "$1" || return 0
+chk_pkginstx "$1" || return 0
+chk_pkgexplicit "$1" || return 0
 $pcmbin -D --asdeps "$1"|trbl_t
-chk_pkgisexplicit "$1" && return 1 || return 0
+chk_pkgexplicit "$1" && return 1 || return 0
 }
 
 manualRemoval(){
 #$1=(pkgs)|$2=vsn (or older) to remove|[$3=new pkgs]
 IFS=" " read -ra mr_pkgo <<< "$1"
-if chk_pkgisinst "${mr_pkgo[0]}" && [[ "$(chk_pkgvsndiff "${mr_pkgo[0]}" "$2")" -le 0 ]]; then
+if chk_pkginstx "${mr_pkgo[0]}" && [[ "$(chk_pkgvsndiff "${mr_pkgo[0]}" "$2")" -le 0 ]]; then
     trbl "attempting manual package removal/replacement of $1..."
     if [[ ! "$3" = "" ]]; then
         # shellcheck disable=SC2086
         $pcmbin -Sw --noconfirm $3 $sf_ignore 2>&1|trbl_t
         if [[ ! "${PIPESTATUS[0]}" = "0" ]]; then trbl "$co_r failed to download $3"; return 1; fi
     fi
-    for p in "${mr_pkgo[@]}"; do chk_pkgisinst "$p" && $pcmbin -Rdd --noconfirm "$p" 2>&1|trbl_t; done
+    for p in "${mr_pkgo[@]}"; do chk_pkginstx "$p" && $pcmbin -Rdd --noconfirm "$p" 2>&1|trbl_t; done
     # shellcheck disable=SC2086
     [[ "$3" = "" ]] || $pcmbin -S --noconfirm $3 $sf_ignore 2>&1|trbl_t
     if [[ ! "${PIPESTATUS[0]}" = "0" ]]; then trbl "$co_r failed to replace $1 with $3"; return 1; fi
@@ -681,7 +690,7 @@ if [ "${conf_a[notify_1enable_bool]}" = "$ctrue" ]; then
         if notify-desktop --help >/dev/null 2>&1; then
             noti_gdbus=$false; noti_send=$false
             trblqin "notify-desktop found, using for notifications"
-        elif chk_pkgisinst "plasma-desktop"; then
+        elif chk_pkginst "plasma-desktop"; then
             if notify-send --help >/dev/null 2>&1; then
                 noti_gdbus=$false; noti_desk=$false
                 trblqin "$co_y KDE Plasma desktop found, falling back to legacy. Please install notify-desktop-git for fully-supported notifications in KDE Plasma"
@@ -831,7 +840,7 @@ fi
 
 if perst_isneeded "${conf_a[update_keys_freq]}" "${perst_a[keys_up_date]}"; then
 while : ; do
-    if chk_pkgisinst "archlinux-keyring" && [[ "$(chk_pkgvsndiff "archlinux-keyring" "20220713-1")" -lt 0 ]]; then break; fi
+    if chk_pkginstx "archlinux-keyring" && [[ "$(chk_pkgvsndiff "archlinux-keyring" "20220713-1")" -lt 0 ]]; then break; fi
         trbl "Refreshing keys..."; pacman-key --refresh-keys  2>&1|trbl_t |tee /dev/tty |grep "Total number processed:" >/dev/null
         err_keys=("${PIPESTATUS[@]}"); if [[ "${err_keys[0]}" -eq 0 ]] || [[ "${err_keys[3]}" -eq 0 ]]; then
             perst_update "keys_up_date"; err[keys]=0; else err[keys]=${err_keys[0]}; trbl "$co_y pacman-key exited with code ${err[keys]}"; fi
@@ -844,7 +853,7 @@ while : ; do
 if ! chk_freespace_all; then err[repo]=1; err_crit="repo"; break; fi
 
 #Does not support installs with xproto<=7.0.31-1
-if chk_pkgisinst "xproto" && [[ "$(chk_pkgvsndiff "xproto" "7.0.31-1")" -le 0 ]]; then
+if chk_pkginstx "xproto" && [[ "$(chk_pkgvsndiff "xproto" "7.0.31-1")" -le 0 ]]; then
     trbl "$co_r old xproto installed - system too old for script to update"; err[repo]=1; err_crit="repo"; break; fi
 
 
@@ -867,7 +876,7 @@ if [[ "${conf_a[repair_1enable_bool]}" = "$ctrue" ]] && [[ "${conf_a[repair_manu
         trbl "Old pacman detected, attempting to use pacman-static..."
         pacman -Sw --noconfirm pacman-static 2>&1|trbl_t
         pacman -U --noconfirm "$(get_pkgfilename "pacman-static")" 2>&1|trbl_t && pcmbin="pacman-static"
-        if ! chk_pkgisinst "pacman-static"; then
+        if ! chk_pkginst "pacman-static"; then
             if dl_verify "pacmanstatic" "$url_repo/master/external/hash_pacman-static" "$url_repo/master/external/pacman-static"; then
                 chmod +x /tmp/xs-autmp-pacmanstatic/pacman-static && pcmbin="/tmp/xs-autmp-pacmanstatic/pacman-static"; fi; fi
         if echo "$pcmbin"|grep "pacman-static" >/dev/null 2>&1 && $pcmbin --help >/dev/null 2>&1; then
@@ -910,7 +919,7 @@ if [[ "${conf_a[repair_1enable_bool]}" = "$ctrue" ]] && [[ "${conf_a[repair_manu
     manualRemoval "wxgtk2" "3.0.5.1-3" #Removed from arch repos 2022-07-14
     manualRemoval "pipewire-media-session" "1:0.4.1-1" "wireplumber" #Replaced 2022-05-10 (bump version when demoted)
     manualRemoval "qpdfview" "0.4.18-1" "evince" #Moved to AUR 2022-04-01 (bump version when updated)
-    if chk_pkgisinst "kvantum-manjaro" && [[ "$(chk_pkgvsndiff "kvantum-manjaro" "0.13.5-1")" -le 0 ]]; then
+    if chk_pkginstx "kvantum-manjaro" && [[ "$(chk_pkgvsndiff "kvantum-manjaro" "0.13.5-1")" -le 0 ]]; then
         for t in adapta-black-breath-theme adapta-black-maia-theme adapta-breath-theme adapta-gtk-theme adapta-maia-theme arc-themes-maia \
             arc-themes-breath matcha-gtk-theme; do manualExplicit "$t"; done; fi #removed from depends of kvmantum-manjaro 2022/02/23
     manualRemoval "galculator-gtk2" "2.1.4-5" "galculator" #Removed from galculator 2021/11/13
@@ -922,7 +931,7 @@ if [[ "${conf_a[repair_1enable_bool]}" = "$ctrue" ]] && [[ "${conf_a[repair_manu
     manualRemoval "knetattach" "5.20.5-1" #Merged into plasma-desktop 2021/01/09
     manualRemoval "gksu-polkit" "0.0.3-2" "zensu" #Removed from manjaro repos 2020/10
     manualRemoval "breeze-kde4" "5.13.4-1"; manualRemoval "oxygen-kde4" "5.13.4-1"; manualRemoval "sni-qt" "0.2.6-5" #removed from repos 2019/05
-    if chk_pkgisinst "phonon-qt4" && [[ "$(chk_pkgvsndiff "phonon-qt4" "4.10.3-1")" -le 0 ]]; then
+    if chk_pkginstx "phonon-qt4" && [[ "$(chk_pkgvsndiff "phonon-qt4" "4.10.3-1")" -le 0 ]]; then
         for t in phonon-qt4-gstreamer phonon-qt4-vlc phonon-qt4-mplayer-git; do manualDepend "$t"; done; fi #these should be depends of phonon-qt4, moved to AUR 2019/05
     manualRemoval "ms-office-online" "20.1.0-1" #moved to AUR 2020/06
     manualRemoval "pyqt5-common" "5.13.2-1" #Removed from repos early 2019/12
@@ -934,7 +943,7 @@ fi
 
 trbl "Updating system packages..."
 for p in $(pacman -Sl | grep "\[installed"|grep -oP "[^ ]*\-system$") ${conf_a[main_systempkgs_str]}; do
-    chk_pkgisinst "$p" && $pcmbin -S --needed --noconfirm "$p" 2>&1|trbl_t
+    chk_pkginstx "$p" && $pcmbin -S --needed --noconfirm "$p" 2>&1|trbl_t
     ((err[sys]+=PIPESTATUS[0])); done
 if [[ ${err[sys]} -ne 0 ]]; then trbl "$co_y system packages failed to update - err:${err[sys]}"; fi
 
@@ -989,11 +998,8 @@ done
 
 #check if AUR pkgs need rebuild
 if [[ "${conf_a[repair_1enable_bool]}" = "$ctrue" ]] && [[ "${conf_a[repair_aurrbld_bool]}" = "$ctrue" ]]; then
-    if ! chk_pkgisinst "rebuild-detector"; then
-        trbl "AUR Helper installed and enabled, and rebuilds are enabled. Installing missing dependency: rebuild-detector"
-        $pcmbin -S --needed --noconfirm rebuild-detector 2>&1|trbl_t
-    fi
-    if chk_pkgisinst "rebuild-detector"; then
+    chk_pkginst "rebuild-detector" || inst_misspkg "rebuild-detector" "AUR Helper installed and enabled, and rebuilds are enabled"
+    if chk_pkginst "rebuild-detector"; then
         trbl "Checking if AUR packages need rebuild..."
         for pkg in $(printf "%s\n" "${!perst_a[@]}"|grep -E "^zrbld:"); do
             if perst_isneeded "${conf_a[repair_aurrbldfail_freq]}" "${perst_a[$pkg]}"; then perst_reset "$pkg"; continue; fi
@@ -1021,7 +1027,7 @@ if [ "${hlpr_a[pikaur]}" = "1" ]; then while :; do
     trbl "Checking if pikaur functional..."
     if testpikaur; then break; fi
     hlpr_a[pikaur]=0; trbl "$co_y AURHelper: pikaur not functioning"
-    if [[ "${conf_a[repair_1enable_bool]}" = "$ctrue" ]] && [[ "${conf_a[repair_pikaur01_bool]}" = "$ctrue" ]] && $pcmbin -Q pikaur >/dev/null 2>&1; then
+    if [[ "${conf_a[repair_1enable_bool]}" = "$ctrue" ]] && [[ "${conf_a[repair_pikaur01_bool]}" = "$ctrue" ]] && chk_pkginst "pikaur"; then
         trblm "Attempting to re-install ${pikpkg:-pikaur}..."
         mkdir "/tmp/xs-autmp-2delete"; if pushd "/tmp/xs-autmp-2delete"; then
             git clone https://github.com/actionless/pikaur.git; if cd pikaur; then
@@ -1040,15 +1046,9 @@ if [[ "${conf_a[aur_1helper_str]}" = "auto" ]]; then
 
 
 #Install KDE notifier dependency (if auto|desk on KDE)
-if [ "${conf_a[notify_1enable_bool]}" = "$ctrue" ]; then 
-    if echo "${conf_a[notify_function_str]}"|grep "auto\|desk" >/dev/null; then
-        if $pcmbin -Q plasma-desktop >/dev/null 2>&1; then
-            if ! $pcmbin -Q notify-desktop-git; then
-                if [ "${hlpr_a[pikaur]}" = "1" ]; then pikaur -S --needed --noconfirm notify-desktop-git; fi
-                if [ "${hlpr_a[apacman]}" = "1" ]; then apacman -S --needed --noconfirm notify-desktop-git; fi
-            fi
-        fi
-    fi
+if [ "${conf_a[notify_1enable_bool]}" = "$ctrue" ] && echo "${conf_a[notify_function_str]}"|grep "auto\|desk" >/dev/null &&\
+chk_pkginst "plasma-desktop" && ! chk_pkginst "notify-desktop-git" && [[ "$((hlpr_a[pikaur]+hlpr_a[apacman]))" -gt "0" ]]; then
+    inst_misspkg "notify-desktop-git" "Notifications enabled and KDE detected"
 fi
 
 
@@ -1082,7 +1082,7 @@ if [[ "${hlpr_a[pikaur]}" = "1" ]]; then
         trbl "Updating AUR packages with custom flags [pikaur]..."
         for i in $(printf "%s\n" "${!flag_a[@]}"); do
             for j in $(echo "$i" | tr ',' ' '); do
-                $pcmbin -Q "$j" >/dev/null 2>&1 && custpkg+=" $j"; done
+                chk_pkginstx "$j" && custpkg+=" $j"; done
             if [[ ! "$custpkg" = "" ]]; then
                 if test_online; then
                     trblm "Updating: $custpkg"
